distributed_llm_trainer/requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
wandb>=0.15.0
pyyaml>=6.0
pytest>=7.3.1